reorganize the file structure to pages and components
