+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Portal from './component/Portal';
-import Render from './component/Render'
+import Portal from './pages/Portal';
+import Render from './pages/Render'
 import Footer from './component/Footer';
