+++ client/src/pages/Home/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import './style.css';
+
+function Home() {
+
+    return (
+        <div className='home'>
+            
+        </div>
+    )
+
+}
+
+export default Home;
